# [Core] FilePath no longer creates extra strings on IsChildPathof.