@Therzok 2016-09-22
The tests were done on test strings:
`asdf`, `asdf\\`
and
`asdf\\test`