@Therzok 2016-09-22
Linux failure unrelated, looks like a bug in gtk# in mono 4.4.